# OpenApiSpec.yaml
openapi: 3.0.0
info:
  title: Rocket Notes Api
  description: Api for remote note management
  version: 0.1.0

paths:
  /note/read/{title}:
    get:
      summary: Returns the content of a single note
      parameters:
        - name: title
          in: path
          required: true
          description: Title of the note
          schema:
            type: string
      responses:
        '200':
          description: The content of the requested note
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  content:
                    type: string
        '404':
          description: The note with the given title does not exist
  /note/listAll:
    get:
      summary: Lists the titles of all available nodes
      responses:
        '200':
          description: An array of all available nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                    content:
                      type: string
  /note/delete/{title}:
    delete:
      summary: Deletes the note with the given title
      parameters:
        - name: title
          in: path
          required: true
          description: Title of the note to delete
          schema:
            type: string
      responses:
        '200':
          description: The note was deleted
        '404':
          description: The note with the given title does not exist
  /note/create:
    post:
      summary: Creates a new note
      requestBody:
        description: The content of the new note
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: 
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: The note was created successfully
        '409':
          description: A note with the same title already exists